+++ lib/spectre/mixin.rb
@@ -1,4 +1,5 @@
 require 'ostruct'
+require 'spectre/logger'
 
 module Spectre
   module Mixin
@@ -11,29 +12,21 @@ module Spectre
 
       def run desc, with: []
         raise "no mixin with desc '#{desc}' defined" unless @@mixins.key? desc
-        Logger.log_debug "running mixin '#{desc}'"
+        Spectre::Logger.log_debug "running mixin '#{desc}'"
 
-        if with.is_a? Array
-          @@mixins[desc].call *with
+        params = with || {}
+
+        if params.is_a? Array
+          @@mixins[desc].call *params
+        elsif params.is_a? Hash
+          @@mixins[desc].call OpenStruct.new(params)
         else
-          @@mixins[desc].call with
+          @@mixins[desc].call params
         end
       end
 
       alias_method :also, :run
       alias_method :step, :run
-    end
-
-    Spectre.register do |config|
-      if not config.key? 'mixin_patterns'
-        return
-      end
-
-      config['mixin_patterns'].each do |pattern|
-        Dir.glob(pattern).each do|f|
-          require_relative File.join(Dir.pwd, f)
-        end
-      end
     end
 
     Spectre.delegate :mixin, :run, :also, :step, to: Mixin